Move input types into separate files
