+++ lighthouse-protocol/src/input/event_source.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+/// An identifier that is unique per client + device combo.
+#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone)]
+pub enum EventSource {
+    String(String),
+    Int(i32),
+}
