+++ lighthouse-protocol/src/input/gamepad_event.rs
@@ -1,0 +1,11 @@
+use serde::{Deserialize, Serialize};
+
+use super::EventSource;
+
+/// A gamepad/controller event.
+#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone)]
+pub struct GamepadEvent {
+    /// The client identifier. Also unique per gamepad.
+    pub source: EventSource,
+    // TODO: Add remaining
+}
