+++ lighthouse-protocol/src/input/key_event.rs
@@ -1,0 +1,14 @@
+use serde::{Deserialize, Serialize};
+
+use super::EventSource;
+
+/// A keyboard event.
+#[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone)]
+pub struct KeyEvent {
+    /// The client identifier.
+    pub source: EventSource,
+    /// Whether the key was pressed.
+    pub down: bool,
+    /// The key pressed, see the docs on JS's `KeyboardEvent.key` for details.
+    pub key: String, // TODO: Extract stronger `Key` type
+}
