+++ lighthouse-protocol/src/input/mod.rs
@@ -1,5 +1,15 @@
+mod event_source;
+mod gamepad_event;
 mod input_event;
+mod key_event;
 mod legacy_input_event;
+mod mouse_button;
+mod mouse_event;
 
+pub use event_source::*;
+pub use gamepad_event::*;
 pub use input_event::*;
+pub use key_event::*;
 pub use legacy_input_event::*;
+pub use mouse_button::*;
+pub use mouse_event::*;
